refactor(plagiarism): update client models, imports, and tests for student answer check
